+++ t/01-pdk/04-request/00-phase_checks.t
@@ -306,6 +306,18 @@ qq{
                 body_filter   = false,
                 log           = false,
                 admin_api     = true,
+            }, {
+                method        = "get_start_time",
+                args          = {},
+                init_worker   = "forced false",
+                certificate   = "pending",
+                rewrite       = true,
+                access        = true,
+                header_filter = true,
+                response      = true,
+                body_filter   = true,
+                log           = true,
+                admin_api     = true,
             },
         }